+++ fourth_Engine/include/Tools/IBLReflectionCapture.h
@@ -23,9 +23,9 @@ namespace fth::ibl
 			return instance;
 		}
 		void Init();
-	    void GenerateCubemapDiffuseIrradiance(const Texture& src, std::string_view filePath, uint32_t width, uint32_t height,  uint32_t samples, Texture& dst);
-	    void GenerateCubemapSpecularIrradiance(const Texture& src, std::string_view filePath, uint32_t width, uint32_t height, uint32_t samples, Texture& dst);
-	    void GenerateGGXSpecularReflectanceLUT(uint32_t width, std::string_view filePath, uint32_t height, uint32_t samples, Texture& dst);
+	    void GenerateCubemapDiffuseIrradiance(const Texture& src, std::string_view filePath, uint16_t width, uint16_t height,  uint32_t samples, Texture& dst);
+	    void GenerateCubemapSpecularIrradiance(const Texture& src, std::string_view filePath, uint16_t width, uint16_t height, uint32_t samples, Texture& dst);
+	    void GenerateGGXSpecularReflectanceLUT(uint16_t width, std::string_view filePath, uint16_t height, uint32_t samples, Texture& dst);
 
 	private:
 		struct UniformData
@@ -35,7 +35,7 @@ namespace fth::ibl
 			float    pad[2];
 		};
 
-		void Capture(const Texture& src, DXGI_FORMAT format, uint32_t width, uint32_t height, uint16_t mips, uint32_t samples, const renderer::PixelShader& targetPS, Texture& dst);
+		void Capture(const Texture& src, DXGI_FORMAT format, uint16_t width, uint16_t height, uint16_t mips, uint32_t samples, const renderer::PixelShader& targetPS, Texture& dst);
 
 		renderer::VertexShader     m_cubemapGenVS;
 		renderer::PixelShader      m_diffuseIrradiancePS;
